+++ src/__tests__/__mocks__/api/handlers.js
@@ -1,0 +1,12 @@
+import { rest } from "msw";
+import { items as databases } from "../responses/databases";
+
+export const handlers = [
+  // Databases
+  rest.get(
+    `${process.env.REACT_APP_BACKEND_URL}/databases`,
+    (req, res, ctx) => {
+      return res(ctx.json(databases));
+    }
+  )
+];