+++ src/__tests__/__mocks__/responses/databases.js
@@ -1,0 +1,47 @@
+export const items = [
+  {
+    id: 1,
+    database_name: "Mongo DB",
+    view_order: 1,
+    use_percentage: 10,
+    created_at: "2023-01-29T19:23:54.000000Z",
+    updated_at: "2023-01-29T19:23:54.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 2,
+    database_name: "MySql",
+    view_order: 2,
+    use_percentage: 30,
+    created_at: "2023-01-29T19:24:09.000000Z",
+    updated_at: "2023-01-29T19:24:09.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 3,
+    database_name: "Maria DB",
+    view_order: 3,
+    use_percentage: 20,
+    created_at: "2023-01-29T19:24:21.000000Z",
+    updated_at: "2023-01-29T19:24:21.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 4,
+    database_name: "PostgreSql",
+    view_order: 4,
+    use_percentage: 20,
+    created_at: "2023-01-29T19:24:40.000000Z",
+    updated_at: "2023-01-29T19:24:40.000000Z",
+    deleted_at: null
+  },
+  {
+    id: 5,
+    database_name: "Oracle",
+    view_order: 5,
+    use_percentage: 20,
+    created_at: "2023-01-29T19:24:56.000000Z",
+    updated_at: "2023-01-29T19:24:56.000000Z",
+    deleted_at: null
+  }
+];