+++ src/views/App.jsx
@@ -1,14 +1,17 @@
 import React from "react";
 import { Col, Container, Row } from "react-bootstrap";
-import {Chart} from "../components/Chart";
-import {FooterPage} from "../components/FooterPage";
-import {items as programmingLanguageOptions} from "../assets/json/programmingLanguageOptions";
-import {items as databasesOptions} from "../assets/json/databasesOptions";
-import {items as workingModeOptions} from "../assets/json/workingModeOptions";
-import {items as frameworksLibrariesOptions} from "../assets/json/frameworksLibrariesOptions";
-import {items as toolsOptions} from "../assets/json/toolsOptions";
+import { Chart } from "../components/Chart";
+import { FooterPage } from "../components/FooterPage";
+import { items as programmingLanguageOptions } from "../assets/json/programmingLanguageOptions";
+import { items as workingModeOptions } from "../assets/json/workingModeOptions";
+import { items as frameworksLibrariesOptions } from "../assets/json/frameworksLibrariesOptions";
+import { items as toolsOptions } from "../assets/json/toolsOptions";
+import { useGetDatabasesQuery } from "../features/apiSlice";
+import databasesOptions from "../utils/methods/databasesOptions";
 
 export function App() {
+  
+  const { data: databases } = useGetDatabasesQuery();
 
   const options = [
     {
@@ -25,7 +28,7 @@ export function App() {
     },
     {
       key: 4,
-      value: databasesOptions
+      value: databases ? databasesOptions(databases) : null
     },
     {
       key: 5,
@@ -40,21 +43,22 @@ export function App() {
           <h1>Alessandro A. resume</h1>
         </Col>
       </Row>
-      {
-        options.map(el => (
-          <Row key={el.key}>
-            <Col>
-              <Chart options={el.value} />
-            </Col>
-          </Row>
-        ))
-      }
+      {options.map(
+        (el) =>
+          el.value && (
+            <Row key={el.key}>
+              <Col>
+                <Chart options={el.value} />
+              </Col>
+            </Row>
+          )
+      )}
       <Row>
         <Col>
           <FooterPage />
         </Col>
       </Row>
-    </Container >
+    </Container>
   );
 }